+++ backend/src/main/java/site/coduo/todo/exception/TodoException.java
@@ -1,4 +1,4 @@
-package site.coduo.todo.domain.exception;
+package site.coduo.todo.exception;
 
 public class TodoException extends RuntimeException {
